Align environment variables between consoles (#30125)
* change to make authServerUrl the same as authUrl

fixes: #29641

* Remove `authUrl` entirely


* Remove file that is unrelated


* Split out and align environment variables between consoles


* Restore removed variables to preserve backwards compatibility


* Also deprecate the `authUrl` for the Admin Console


---------
