+++ js/apps/admin-ui/src/Root.tsx
@@ -1,5 +1,7 @@
-import { KeycloakProvider, environment } from "@keycloak/keycloak-ui-shared";
+import { KeycloakProvider } from "@keycloak/keycloak-ui-shared";
+
 import { App } from "./App";
+import { environment } from "./environment";
 
 export const Root = () => (
   <KeycloakProvider environment={environment}>
